Move StyleManager to standalone file There are many file changes because they imported the makeStyleProcessor, which name changed from makeStyleManager
