+++ index.html
@@ -2,6 +2,6 @@
 <!-- Mock page for testing -->
 <html>
     <body>
-        <script type="module" src="makeStyleManager.test.mjs"></script>
+        <script type="module" src="tests/makeStyleManager.test.mjs"></script>
     </body>
 </html>
